Replace font subtabs with more... icons
